fix(cli): Prevent no such file error with optional `.env` on build
https://github.com/ecomplus/store/runs/7664420030\?check_suite_focus\=true

+++ packages/cli/src/build.ts
@@ -8,11 +8,14 @@ const copyFunctionsConfig = async () => {
   for (let i = 0; i < dirents.length; i++) {
     if (dirents[i].isDirectory() && dirents[i].name.charAt(0) !== '.') {
       for (let ii = 0; ii < filesToCopy.length; ii++) {
-        // eslint-disable-next-line no-await-in-loop
-        await fs.copy(
-          path.join(functionsDir, filesToCopy[ii]),
-          path.join(functionsDir, dirents[i].name, filesToCopy[ii]),
-        );
+        const srcPath = path.join(functionsDir, filesToCopy[ii]);
+        if (fs.existsSync(srcPath)) {
+          // eslint-disable-next-line no-await-in-loop
+          await fs.copy(
+            srcPath,
+            path.join(functionsDir, dirents[i].name, filesToCopy[ii]),
+          );
+        }
       }
     }
   }
